Create task-2 Student grade calculator
In this program, the user inputs their marks for each subject (out of 100). Then, the program calculates the total marks by summing up the marks obtained in all subjects. The average percentage is calculated by dividing the total marks by the total number of subjects. Based on this average percentage, the program assigns grades according to predetermined criteria. Finally, it displays the total marks, average percentage, and the corresponding grade to the user.