Rename Task 1 - Number Game to Task 1-Number Game
In essence, it's a number guessing game where the user tries to guess a randomly generated number within a specified range. The user gets feedback on their guesses and has a limited number of attempts to get it right. They can play multiple rounds, and their performance is tracked, possibly through a score based on the number of correct guesses.